+++ index.js
@@ -7,13 +7,6 @@
 'use strict'
 
 /**
- * Simple expression to split token list.
- * @private
- */
-
-var TOKEN_LIST_REGEXP = / *, */
-
-/**
  * Module exports.
  * @public
  */
@@ -50,13 +43,34 @@ function forwarded (req) {
  */
 
 function parse (header) {
-  if (!header) {
-    return []
+  var end = header.length
+  var list = []
+  var start = header.length
+
+  // gather addresses, backwards
+  for (var i = header.length - 1; i >= 0; i--) {
+    switch (header.charCodeAt(i)) {
+      case 0x20: /*   */
+        if (start === end) {
+          start = end = i
+        }
+        break
+      case 0x2c: /* , */
+        if (start !== end) {
+          list.push(header.substring(start, end))
+        }
+        start = end = i
+        break
+      default:
+        start = i
+        break
+    }
   }
 
-  return header
-    .trim()
-    .split(TOKEN_LIST_REGEXP)
-    .filter(Boolean)
-    .reverse()
+  // final address
+  if (start !== end) {
+    list.push(header.substring(start, end))
+  }
+
+  return list
 }