+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * perf: improve header parsing
   * perf: reduce overhead when no `X-Forwarded-For` header
 
 0.1.1 / 2017-09-10